vastly simplify chat history merging
